Mais alguns ajustes
foi alterado algumas coisas no banco de dados, atualize o banco

+++ ISports/Models/Organizador.cs
@@ -8,7 +8,7 @@ namespace ISports.Models
     public class Organizador : Usuario
     {
         public int Permissao { get; set; }
-        public double Qualificacao { get; set; }
+        public decimal Qualificacao { get; set; }
 
     }
 }
